backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Jolkkonen, Johannes (External) <johannes.jolkkonen@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.30.5"
python-decouple = "^3.8"
instructor = "^1.3.2"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
qdrant-client = "^1.9.1"
hubspot-api-client = "^9.0.0"
requests = "^2.32.3"
pandas = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
